Move artifact verification to the ingester, remove signature_verification and github_workflow from the ArtifactVersion protobuf
In order to support richer provenance checks in our profiles, we need to
move the artifact provenance evaluation to the ingestor from the current
model where the artifact signatures are evaluated at the point the
artifacts are fetched - either during repo register, profile create or
on receiving a webhook message.

Since the ingestor won't be storing the signature_verification and
github_workflow attributes anymore to the database, let's just remove
those.

Co-authored-by: Radoslav Dimitrov <radoslav@stacklok.com>