Define the protobuf API for selectors
Creates a new protobuf message SelectorEntity that represents a set of
attributes for a typed entity to be used in profile selectors.

At the moment, onyl SelectorRepository and SelectorArtifact are used and
implemented. Additionally, the key-value store is not implemented either
as it requires a new set of provider interfaces to fill this store.

Fixes: 3757