Add ProviderID and ProjectID as ByUpstreamHint parameters
Adds two new parameters to the ByUpstreamHint structure. These function
a little differently than the ones previously in the sense that they are
not hints per se, but rather used directly as parameters to the database
lookup.

The reasoning is that when refreshing an entity, we often know which
entity to refresh exactly (we might have just created it or we are
iterating over entities), but at the same time, it doesn't make much
sense to add yet another API to the property service.